backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
python-socketio==5.10.0
FastAPI==0.104.1
uvicorn==0.24.0
python-nmap==0.7.1
SQLAlchemy==2.0.23
alembic==1.12.1
redis==5.0.1
celery==5.3.4
schedule==1.2.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
websockets==12.0
aiofiles==23.2.1
asyncio==3.4.3
python-dotenv==1.0.0
requests==2.31.0
reportlab==4.0.7
matplotlib==3.8.2
seaborn==0.13.0

